This is a large update:
I have redone the hole backend of the softmod, now the shadowc img and partition are all created on the Xbox. I also patch UnleashX on the xbox to access the protected C partition, doing it this way allows for a few benifits.

1) Softreset from NKPatcher Settings back to your dashboard.
2) Allows the virtual drive to disable the shadowc.
3) ShadowC is never disabled.

I have also cleaned up the and rebranded the EEPROM Backeruper to 2016 Softmod Tool Kit, as this now does a lot more than just backing up the eeprom.

Also cleaned up NKPatcher when its looking for the video.bin it now matches the rest of the runtime configs.

Updated Chimp to the latest revision.
( New Chimp loader and it now locks your HDD upon load of the Chimp menu and then unlocks it, if it fails to lock the drive it will notify you that the drive cannot be locked.
So this is a sure way to tell if your HDD is lockable, as I try to lock it ;) and if it fails well its not lockable. )

Removed the no sc disc folder and its content as its no longer needed.

all round clean up of files and code.