Updated the 2016 Softmod Tool Kit to change the serial number on the virtual eeprom when enabled to "V_EEPROM ENB" This way you know if its enabled or not :)
Softmod site of things

Fixed the list high for all the skins.
Fixed the main skin files (A) Select background being off a little.
Updated to the "2016 Softmod Tool Kit" tool. Now when Virtual EEPROM is enabled it will now display "V-EEPROM ENB" as the systems serial number.
Updated the Virtual EEPROM settings menu:
	> You now cannot run the "Enable" option more than once, as the virtual eeprom is write protected until disabled and the system had restarted.
	> Disable now has a mandatory restart.

Also added a new batch file to build the XISO without the games folder as that puts the IOS to over 1GB.
( I don't want that for testing on the Xbox )