Changed the Upgrade and Update softmod options on the extras disc to only use the E:\Prep directory. ( moving it back to here as I have a pop up that states you need a minimum of 30MB on E )
Changed how the 5960 modded dashboard works, if persistent softmod is disabled it will show the language menu and if its enable it will hide it.

Removed the rar compression from best to default. ( this may fix the rare case of it failing to extract )


Other source side edits, basically a cleanup.